# Improve performance when no transformations are necessary