@meowsus 2018-04-26
> However, this breaks some tests since the resulting SVG code is taken directly from the asset file and is no longer returned in the beautifully consistent format that Nokogiri provides.

If I understand this correctly, and for what it matters, if the argument is code beautification over performance I would argue in favor of performance.