@kylerippey 2018-04-26
> If I understand this correctly, and for what it matters, if the argument is code beautification over performance I would argue in favor of performance.

Agreed. My chief concern with the change that I was experimenting with is that it would change the behavior of the method slightly. In my very brief test I tried adding a short circuit into the `TransformPipeline.generate_html_from` method, like so, to avoid parsing the input with Nokogiri if there are no transformations to be done:
```ruby
def self.generate_html_from(svg_file, transform_params)
  return svg_file.to_s unless transform_params.any? || Transformations.custom_transformations.any?
  ...
```

However, it alters the formatting of returned SVGs slightly because the markup is no longer cleaned up by Nokogiri, as demonstrated by this failing spec:
```
InlineSvg::ActionView::Helpers#inline_svg when passed IO object argument return valid svg for file
     Failure/Error: expect(output).to eq "<svg xmlns=\"http://www.w3.org/2000/svg\" xml:lang=\"en\" role=\"presentation\"><!-- This is a test comment --></svg>\n"

       expected: "<svg xmlns=\"http://www.w3.org/2000/svg\" xml:lang=\"en\" role=\"presentation\"><!-- This is a test comment --></svg>\n"
            got: "<svg xmlns=\"http://www.w3.org/2000/svg\"\n     xml:lang=\"en\"\n     role=\"presentation\"><!-- This is a test comment --></svg>\n"
```

This represents a change in behavior that some users of the gem may not be expecting. Also, the markup returned when you pass transformation options may be slightly different than when you do not pass options. It's a minor inconsistency, and I don't think it's a _huge_ issue, but I figured it was worth discussing before throwing it into a PR. 😄