@jamesmartin 2018-04-26
@kylerippey first of all, thanks for taking the time to open this PR. ❤️  Based on the changes and your benchmarks it looks like a no-brainer. 👍

---

> If we skip parsing SVG icons with Nokogiri entirely when there are no transformations to be applied, we could speed up the basic use case by ~90x!

> If I understand this correctly, and for what it matters, if the argument is code beautification over performance I would argue in favor of performance.

I agree with you and @meowsus here: In principle, I'm also in favor of performance over code beautification. 

> However, it alters the returned SVG markup slightly because it's no longer cleaned up by Nokogiri
> ...
> This represents a change in behavior that some users of the gem may not be expecting. Also, the markup returned when you pass transformation options may be slightly different than when you do not pass options. It's a minor inconsistency, and I don't think it's a huge issue, but I figured it was worth discussing before throwing it into a PR.

Thanks for starting this discussion. I'm definitely in favor of trying this out. I think if we're being true to semver, it would warrant a major version bump. As there are no major feature requests outstanding I'd be in favor of starting a v2.0 branch that is solely focused on performance and including more of your ideas for improvement. What do you think?