@kylerippey 2018-04-26
> I'm definitely in favor of trying this out. I think if we're being true to semver, it would warrant a major version bump. As there are no major feature requests outstanding I'd be in favor of starting a v2.0 branch that is solely focused on performance and including more of your ideas for improvement. What do you think?

I think that would be fantastic. I'd really love to figure out some ways to speed up the more common use cases, such as adding classes to SVGs. Being able to potentially alter the API might give us more opportunities to consider larger architectural changes. 👍